feat: add dependencies, configute typescript and linters
